添加检查更新按钮
