+++ front/src/i18n/zh-CN.js
@@ -53,7 +53,8 @@ export default {
   },
   extension: {
     conditions: '使用须知',
-    conditionsContent: '首次使用扩展模块时，必须安装由Proxyee Down随机生成的一个CA证书，点击下面的安装按钮并按系统的引导进行确认安装。(注意：程序会在安装前检测操作系统中是否有安装过证书，当检测到有安装的情况会提示删除对应的旧CA证书)',
+    conditionsContent:
+      '首次使用扩展模块时，必须安装由Proxyee Down随机生成的一个CA证书，点击下面的安装按钮并按系统的引导进行确认安装。(注意：程序会在安装前检测操作系统中是否有安装过证书，当检测到有安装的情况会提示删除对应的旧CA证书)',
     install: '安装',
     globalProxy: '全局代理',
     proxyTip: '点击查看说明',
@@ -69,6 +70,7 @@ export default {
     actionUpdate: '更新',
     actionInstall: '安装',
     uninstall: '卸载',
+    uninstallTip: '确定卸载此扩展吗？',
     actionDetail: '详情',
     switch: '开关',
     downloadingTip: '下载中...[服务器(',
@@ -118,7 +120,9 @@ export default {
       community: '官方社区：',
       tutorial: '使用教程：',
       feedback: '问题反馈：',
-      currentVersion: '当前版本：'
+      currentVersion: '当前版本：',
+      checkUpdate: '检查更新：',
+      noNewVersion: '已经是最新版本'
     },
     team: {
       title: '团队'
